+++ src/main/java/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -1036,6 +1036,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                     .altIcons("extra-icons/video_alt3.svg"),
                 ofFile("ext_xiph", "extra-icons/xiph.svg", "Xiph audio: *.flac, *.oga, *.ogg, *.opus, *.spx")
                     .end(".flac", ".oga", ".ogg", ".opus", ".spx"),
+                ofFile("ext_yang", "extra-icons/yang.svg", "YANG: *.yang")
+                    .end(".yang"),
 
                 ofFile("ext_libreoffice_calc", "extra-icons/officedocs/localc.png", "LibreOffice Calc: *.ods")
                     .end(".ods")